adding EnhancedDialog to BrowseFilesDialog.tsx
